requirements.txt
alembic == 1.13.1
asyncpg == 0.29.0
bcrypt == 4.1.2
fastapi == 0.108.0
psycopg2-binary == 2.9.9
pydantic[email] == 2.5.3
pydantic-settings == 2.1.0
pyjwt == 2.8.0
pytest == 7.4.4
sqlalchemy == 2.0.25
uvicorn[standard] == 0.25.0
